requirements1.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
contourpy==1.3.2
crccheck==1.3.0
cycler==0.12.1
depthai==2.30.0.0
filelock==3.13.1
fonttools==4.58.1
fsspec==2024.6.1
inputs==0.5
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pillow==11.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
pyvesc @ git+https://github.com/LiamBindle/PyVESC.git@d2f2076afe91dafa920a38a437528819697e05e3
PyYAML==6.0.2
scipy==1.15.3
simsimd==6.4.7
six==1.17.0
stringzilla==3.12.5
sympy==1.13.3
torch==2.7.1+cpu
torchaudio==2.7.1+cpu
torchvision==0.22.1+cpu
typing-inspection==0.4.1
typing_extensions==4.12.2
